Add a logout button in the navbar and improve the login/logout flow
Add a logout button to the navbar. In order for this to work, template
registration/logged_out.html needs to be added to override the one
provided by Django's admin. In order for this overriding to work,
'xorgauth' needs to come before 'django.contrib.admin' in
INSTALLED_APPS.

While at it, translate the login button text.