Make sure Django 1 gets installed instead of Django 2.0
The last release of django-oidc-provider (0.5.2) was not compatible with
Django 2.0. A fix has been integrated upstream
(https://github.com/juanifioren/django-oidc-provider/pull/216) but no
released version of django-oidc-provider comes with it yet.

Wait for such a release by marking Django 2.0 as unsupported for now in
the setup.py requirements, event with Python 3.