Update the groups perms when importing data over an existing member
When importing the accounts data several times into the same database
(in order to keep it roughly in sync with the original database along
the days of beta-testing), group membership permissions can change. Do
not crash the script when this occurs (because (user,group) is unique in
the group membership table), but instead update the permission
accordingly.