Do not use email field if it is empty
When the email address (or username) entered into the password reset
form is empty, "cleaned_data = super(PasswordResetFrom, self).clean()"
drop this field, but PasswordResetFrom.clean() expects
cleaned_data['email'] to be present. Remove such an assumption.

It is currently easy to trigger an HTTP 500 error by entering a space in
the recovery email form field.