Add missing "blank=True" attributes to nullable fields
When editing an authgroupex client entry in Django admin, the last used
field is required, even when a client never used the authentication
endpoint. This is due to a missing blank=True in the model. Add it.

While at it, make User.xorgdb_uid blank too, for the same reason.